Fix memory error causing crash when lines got too long
It is related to the buffer size for the full file, if the buffer has to
double in size, mem location errors occur, for now i've increased the
base memory allocated but this is not the long term solution. Found the
same bug in version 1 though the original buffer is set to a very large
size so this error never triggered. I need time to think about how I
want to proceed with this part of the program